% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cora.R
\name{adjust_BH}
\alias{adjust_BH}
\title{Benjamini-Hochberg (BH) p-value adjustment}
\usage{
adjust_BH(rawp)
}
\arguments{
\item{rawp}{Numeric vector of raw p-values.}
}
\value{
Numeric vector of BH-adjusted p-values.
}
\description{
Standard BH step-up procedure for FDR control under independence or PRDS.
Wrapper around \code{\link[stats]{p.adjust}} from base R.
}
\examples{
rawp <- c(0.001, 0.01, 0.03, 0.5, 0.9)
adjust_BH(rawp)

}
\references{
Benjamini Y, Hochberg Y. Controlling the false discovery rate: a practical
and powerful approach to multiple testing. J R Stat Soc B. 1995;57:289-300.
}
