% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cora.R
\docType{data}
\name{cora}
\alias{cora}
\title{CORA data set}
\format{A data frame with 16 variables: \code{id}, \code{title}, \code{book_title}, \code{authors}, \code{address}, \code{date}, \code{year}, \code{editor},\code{journal}, \code{volume}, \code{pages}, \code{publisher}, \code{institution}, \code{type}, \code{tech}, \code{note}.

This data set includes 1879 CORA research papers.
It is appropriate for performing various types of record linkage and can be assessed by standard record linkage methods.}
\usage{
cora
}
\description{
This provides a record linkage data set with information about different CORA research papers.
}
\examples{
head(cora)
dim(cora)

}
\keyword{datasets}
