% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cora.R
\docType{data}
\name{cora_gold_update}
\alias{cora_gold_update}
\title{Cora Gold Update}
\format{A data frame with 2 variables: \code{cora_id}, \code{unique_id}

This data set includes the matched record pairs based on ID from the CORA data set.
This data set can be used to evaluate the performance of record linkage methods performed on the CORA data set.}
\usage{
cora_gold_update
}
\description{
This data set includes the matched record pairs based on ID.
}
\examples{
head(cora_gold_update)
dim(cora_gold_update)

}
\keyword{datasets}
